feat: Step 23 - phi0-phi7 END-TO-END EPOCH + Phase-0 final audit CLOSED (DEC-216/217, CA-R93..R97): one binary runs a full epoch through every pillar - real-beacon envelopes, sigma_user, ordering lock, order-bound threshold decrypt, the fold over the shuffled order, the 41-byte accumulator, the light-client certificate; seam negatives at every boundary; CA-R88 closed; the honest Phase-0 boundary stated (DEC-217). SUITE 23/23 - PHASE 0 COMPLETE.

+++ conformance/test_step17.cpp
@@ -80,10 +80,10 @@ int main() {
     CHECK(nivc::ACC_SIZE == 41, "G17F accumulator 41B");
 
     // 6. The order matches (M2 ordering → the fold's input)
-    printf("[dbg] count=%u golden=%u\n", result.acc.count, golden::G17F_COUNT);
+    printf("[dbg] count=%llu golden=%llu\n", (unsigned long long)result.acc.count, (unsigned long long)golden::G17F_COUNT);
 
     if (failures) { std::printf("\n%d FAILURE(S) IN STEP 17\n", failures); return 1; }
     std::printf("step17 conformance: ALL GREEN (M2 ordering -> fold -> NIVC close, "
-                "%u entries, 41B accumulator)\n", result.acc.count);
+                "%llu entries, 41B accumulator)\n", (unsigned long long)result.acc.count);
     return 0;
 }
